Revert "refactor: 비회원 ssr로 접근 제어 "
