+++ src/components/menu/navBar/index.tsx
@@ -28,13 +28,6 @@ export interface NavBarProps {
 
 const NavBar = ({ name, isOpen, status }: NavBarProps) => {
   const route = useRouter();
-  const submit = () => {
-    mainInstance.logout().then(() => {
-      removeAuthToken('accessToken');
-      window.localStorage.removeItem('persist');
-      route.push('/');
-    });
-  };
 
   return (
     <NavContainer status={status}>
@@ -90,7 +83,16 @@ const NavBar = ({ name, isOpen, status }: NavBarProps) => {
       </List>
       <Bottom>
         <Bar />
-        <LogoutLi css={{ cursor: 'pointer' }} onClick={submit}>
+        <LogoutLi
+          css={{ cursor: 'pointer' }}
+          onClick={() => {
+            mainInstance.logout().then(() => {
+              removeAuthToken('accessToken');
+              window.localStorage.removeItem('persist');
+              route.push('/');
+            });
+          }}
+        >
           <Logout />
           <span>로그아웃</span>
         </LogoutLi>
