+++ src/pages/create-tape.tsx
@@ -1,28 +1,14 @@
+import SpinnerIcon from 'components/button/spinner';
 import { ButtonLayout } from 'components/button/style';
 import Input from 'components/input';
-import SpinnerIcon from 'components/spinner';
 import TapeSVG from 'components/tape/tape';
 import Title from 'components/title';
 import { MAX_LENGTH } from 'constants/maxTextLen';
 import useInput from 'hooks/useInput';
 import useLoading from 'hooks/useLoading';
-import { NextConfig } from 'next';
 import { useRouter } from 'next/router';
 import { useUserStore } from 'store';
 import { Box, Info, InputBox } from 'styles/create-tape';
-
-export const getServerSideProps = (context: NextConfig) => {
-  const { accessToken } = context.req.cookies;
-
-  if (!accessToken) {
-    return {
-      redirect: {
-        destination: '/',
-        permanent: false,
-      },
-    };
-  }
-};
 
 const CreateTape = () => {
   const { setUserData, date } = useUserStore();
