+++ src/pages/decorate-tape.tsx
@@ -1,10 +1,9 @@
+import SpinnerIcon from 'components/button/spinner';
 import { ButtonLayout } from 'components/button/style';
 import ColorPlate from 'components/colorPlate';
-import SpinnerIcon from 'components/spinner';
 import TapeSVG from 'components/tape/tape';
 import Title from 'components/title';
 import useLoading from 'hooks/useLoading';
-import { NextConfig } from 'next';
 import { useRouter } from 'next/router';
 import { useColorStore, useResponsUserStore, useUserStore } from 'store';
 import { DecoContainer, DecoZone, Middie } from 'styles/decorate-tape';
@@ -12,19 +11,6 @@ import { Color } from 'types';
 import subInstance from 'utils/api/sub';
 
 import theme from '../styles/theme';
-
-export const getServerSideProps = (context: NextConfig) => {
-  const { accessToken } = context.req.cookies;
-
-  if (!accessToken) {
-    return {
-      redirect: {
-        destination: '/',
-        permanent: false,
-      },
-    };
-  }
-};
 
 const DecorateTape = () => {
   const { tapeColor, setTapeColor } = useColorStore();
