+++ src/pages/guest/[id]/guest-entry.tsx
@@ -1,6 +1,6 @@
 'use client';
+import SpinnerIcon from 'components/button/spinner';
 import { ButtonLayout } from 'components/button/style';
-import SpinnerIcon from 'components/spinner';
 import TapeSVG from 'components/tape/tape';
 import useLoading from 'hooks/useLoading';
 import dynamic from 'next/dynamic';
