+++ src/pages/modify-decorate-tape.tsx
@@ -1,10 +1,9 @@
+import SpinnerIcon from 'components/button/spinner';
 import { ButtonLayout } from 'components/button/style';
 import ColorPlate from 'components/colorPlate';
-import SpinnerIcon from 'components/spinner';
 import TapeSVG from 'components/tape/tape';
 import Title from 'components/title';
 import useLoading from 'hooks/useLoading';
-import { NextConfig } from 'next';
 import dynamic from 'next/dynamic';
 import { useRouter } from 'next/router';
 import { useColorStore, useResponsUserStore, useUserStore } from 'store';
@@ -14,19 +13,6 @@ import { Color } from 'types';
 import subInstance from 'utils/api/sub';
 
 const MenuLayout = dynamic(() => import('components/menu'));
-
-export const getServerSideProps = (context: NextConfig) => {
-  const { accessToken } = context.req.cookies;
-
-  if (!accessToken) {
-    return {
-      redirect: {
-        destination: '/',
-        permanent: false,
-      },
-    };
-  }
-};
 
 const ModifyDecorateTape = () => {
   const { tapeColor, setTapeColor } = useColorStore();
