+++ src/pages/withdrawal.tsx
@@ -1,12 +1,11 @@
 import Close from '@icon/close.svg';
+import SpinnerIcon from 'components/button/spinner';
 import CheckBox from 'components/checkBox';
 import Dropdown from 'components/dropDown';
-import SpinnerIcon from 'components/spinner';
 import TapeSvg from 'components/tape/tape';
 import Textarea from 'components/textarea';
 import { WITHDRAWAL } from 'constants/withdrawal';
 import useLoading from 'hooks/useLoading';
-import { NextConfig } from 'next';
 import { useRouter } from 'next/router';
 import { ChangeEvent, useState } from 'react';
 import { dropdownStore, useUserStore } from 'store';
@@ -22,19 +21,6 @@ import {
 import { WithdrawalType } from 'types';
 import mainInstance from 'utils/api/main';
 import { removeAuthToken } from 'utils/storage/authCookie';
-
-export const getServerSideProps = (context: NextConfig) => {
-  const { accessToken } = context.req.cookies;
-
-  if (!accessToken) {
-    return {
-      redirect: {
-        destination: '/',
-        permanent: false,
-      },
-    };
-  }
-};
 
 const Withdrawal = () => {
   const [checked, setChecked] = useState<boolean>(false);
